Added c_matrix3x3::rotation_around_axis. Fixed c_matrix3x3/2x2 rotation function. Fixed i_matrix multiplication operator by column.

+++ src/data-types/matrix/matrix2x2.h
@@ -6,7 +6,7 @@ template <template <typename, typename, typename, size_t, size_t> class major_ty
 class c_matrix2x2 : public null::sdk::i_matrix<major_type_t, float, 2, 2> {
 public: using null::sdk::i_matrix<major_type_t, float, 2, 2>::i_matrix;
 	  static c_matrix2x2 rotation(const radians_t& angle, const math::e_rotation& direction = math::e_rotation::ccw ) {
-		  const float cos{ std::cos(angle) }, sin{ std::sin(angle) };
+		  const float cos{ std::cosf(angle) }, sin{ std::sinf(angle) };
 		  return { {
 			  { cos,												direction == math::e_rotation::ccw ? -sin : sin },
 			  { direction == math::e_rotation::ccw ? sin : -sin,	cos },
